+++ app/src/app/(pages)/addons/layout.tsx
@@ -4,7 +4,7 @@ import { getContentLinks } from '@/lib';
 import '../pages.layout.css';
 import MotionWrapper from '@/components/MotionWrapper';
 import { SidebarProvider } from '@/components/SidebarToggleContext';
-import MobileUI from '@/components/MobileUI';
+import DrawerContainer from '@/components/DrawerContainer';
 
 export default function AddonsLayout({ children }: { children: ReactNode }) {
   const links = getContentLinks('addons');
@@ -18,9 +18,9 @@ export default function AddonsLayout({ children }: { children: ReactNode }) {
 
         {/* content area */}
         <div className="flex-1 relative">
-          <MobileUI links={links} type="addons">
+          <DrawerContainer links={links} type="addons">
             <MotionWrapper>{children}</MotionWrapper>
-          </MobileUI>
+          </DrawerContainer>
         </div>
       </main>
     </SidebarProvider>
